dualserve/django-ml/requirements.txt
django
djangorestframework
gunicorn
onnxruntime
scikit-learn
pytest
pytest-django
numpy 